BF-28: ensure RefreshIndicator works for Gyms Near Me
