+++ lib/ui/home/home_area/articles_card.dart
@@ -1,0 +1,133 @@
+import 'package:auto_size_text/auto_size_text.dart';
+import 'package:bfit_tracker/theme.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+
+class ArticlesCard extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      elevation: 2,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(12.0),
+      ),
+      color: Colors.white,
+      child: Column(
+        children: <Widget>[
+          Expanded(
+            child: Stack(
+              children: <Widget>[
+                Align(
+                  alignment: Alignment.center,
+                  child: Padding(
+                    padding: const EdgeInsets.all(12),
+                    child: SvgPicture.asset(
+                      'assets/images/undraw_healthy_habit.svg',
+                    ),
+                  ),
+                ),
+                Positioned(
+                  top: 2,
+                  right: 12,
+                  child: ButtonTheme(
+                    minWidth: 68,
+                    height: 24,
+                    child: RaisedButton(
+                      padding: const EdgeInsets.all(0),
+                      disabledColor: mainTheme.accentColor,
+                      onPressed: null,
+                      child: AutoSizeText(
+                        '5 min read',
+                        maxLines: 1,
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 8,
+                        ),
+                      ),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(12.0),
+                      ),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+          Align(
+            alignment: Alignment.bottomCenter,
+            child: Container(
+              constraints: const BoxConstraints(
+                minWidth: double.infinity,
+              ),
+              decoration: BoxDecoration(
+                color: mainTheme.primaryColor,
+                borderRadius: BorderRadius.all(
+                  Radius.circular(12.0),
+                ),
+              ),
+              child: Padding(
+                padding: const EdgeInsets.symmetric(
+                  vertical: 6,
+                ),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  children: <Widget>[
+                    Align(
+                      alignment: Alignment.topCenter,
+                      child: AutoSizeText(
+                        '10 Ways To Stay Motivated At The Gym',
+                        maxLines: 2,
+                        minFontSize: 16,
+                        maxFontSize: 32,
+                        style: TextStyle(
+                          fontWeight: FontWeight.bold,
+                          color: Colors.white,
+                        ),
+                      ),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        vertical: 3,
+                      ),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        horizontal: 12,
+                      ),
+                      child: AutoSizeText(
+                        'Find out the best kept secrets from the one and only Nick Mitchell!',
+                        maxLines: 2,
+                        minFontSize: 12,
+                        maxFontSize: 28,
+                        style: TextStyle(
+                          color: Colors.white,
+                        ),
+                      ),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.only(
+                        right: 16,
+                      ),
+                      child: Align(
+                        alignment: Alignment.bottomRight,
+                        child: AutoSizeText(
+                          'Dr. Divyesh Vala',
+                          maxLines: 1,
+                          maxFontSize: 12,
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontStyle: FontStyle.italic,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
