+++ lib/ui/home/home_area/gyms_near_me.dart
@@ -1,9 +1,109 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:bfit_tracker/blocs/gym/gym_bloc.dart';
+import 'package:bfit_tracker/blocs/location/location_bloc.dart';
 import 'package:bfit_tracker/controllers/gym_controller.dart';
 import 'package:bfit_tracker/models/gym.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+
+class NearByGymsCard extends StatefulWidget {
+  NearByGymsCard({Key key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() {
+    return _NearByGymsCardState();
+  }
+}
+
+class _NearByGymsCardState extends State<NearByGymsCard> {
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  Future<void> _refresh() {
+    BlocProvider.of<GymBloc>(context)..add(LoadGym());
+    return Future.value();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      elevation: 2,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(12.0),
+      ),
+      color: Colors.white,
+      child: LayoutBuilder(
+        builder: (context, constraints) => SingleChildScrollView(
+          physics: AlwaysScrollableScrollPhysics(),
+          child: SizedBox(
+            height: constraints.maxHeight,
+            child: Column(
+              children: <Widget>[
+                Padding(
+                  padding: const EdgeInsets.symmetric(
+                    vertical: 12,
+                  ),
+                  child: Container(
+                    child: AutoSizeText(
+                      "Gyms Near Me",
+                      maxLines: 1,
+                      minFontSize: 18,
+                      style: TextStyle(
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ),
+                Expanded(
+                  child: RefreshIndicator(
+                    onRefresh: this._refresh,
+                    child: BlocConsumer<LocationBloc, LocationState>(
+                      listener: (BuildContext context, state) {},
+                      builder: (BuildContext context, state) {
+                        if (state is LocationError) {
+                          // We use a Stack and empty ListView so that
+                          // the RefreshIndicator works
+                          return Stack(
+                            children: <Widget>[
+                              Center(
+                                child: AutoSizeText(
+                                  'BFit Tracker is unable to locate your current position',
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                              ListView()
+                            ],
+                          );
+                        }
+
+                        if (state is Located) {
+                          return GymsNearMe();
+                        } else {
+                          // We use a Stack and empty ListView so that
+                          // the RefreshIndicator works
+                          return Stack(
+                            children: <Widget>[
+                              Center(
+                                child: CircularProgressIndicator(),
+                              ),
+                              ListView()
+                            ],
+                          );
+                        }
+                      },
+                    ),
+                  ),
+                )
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
 
 class GymsNearMe extends StatefulWidget {
   @override
@@ -24,8 +124,18 @@ class _GymsNearMeState extends State<GymsNearMe> {
           future: BlocProvider.of<GymBloc>(context).state.props.first,
           builder: (BuildContext context, AsyncSnapshot snapshot) {
             if (snapshot.hasError) {
-              return Center(
-                child: AutoSizeText('Unable to load gyms near you'),
+              // We use a Stack and empty ListView so that
+              // the RefreshIndicator works
+              return Stack(
+                children: <Widget>[
+                  Center(
+                    child: AutoSizeText(
+                      'Unable to load gyms near you',
+                      textAlign: TextAlign.center,
+                    ),
+                  ),
+                  ListView()
+                ],
               );
             }
