+++ pubspec.yaml
@@ -7,7 +7,7 @@ environment:
   flutter: "1.20.4"
 
 dependency_overrides:
-  flutter_svg: ^0.19.0
+  flutter_svg: ^0.19.1
   meta: ^1.2.3
 
 dependencies:
@@ -35,7 +35,7 @@ dependencies:
   geolocator: ^6.0.0+4
   google_sign_in: ^4.5.3
   health: 1.0.2
-  height_slider: ^1.2.0
+  height_slider: ^1.3.0
   http: ^0.12.2
   intl: ^0.16.1
   jiffy: ^3.0.1
